move checkFresh etc to their own file
