+++ packages/webapp/src/actions/errors.ts
@@ -14,7 +14,7 @@ import { ContractWriteError } from "src/actions/libContractWrite"
 import { GIT_ISSUES } from "src/constants"
 import { setError } from "src/store/actions"
 
-import { StaleError } from "src/store/read"
+import { StaleError } from "src/store/checkFresh"
 import { TimeoutError } from "src/utils/errors"
 import { ProofCancelled, ProofError, ProofTimeoutError } from "src/utils/zkproofs/proofs"
